ENGR00153462 IPUv3 lib:Improve the logic of fitting fb
This patch improves the logic of fitting framebuffer
for show-to-fb case. The desired output resolution
should be aligned with the original one, and the
desired virtual yres should be less than or equal to
the original one, otherwise, we reconfigure
the framebuffer.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>
(cherry picked from commit dab8c562ee24726f988f1acf8f9dfad80610a0be)