ENGR00216013 vpu: add phy address check before start use.
this patch is adding a protection for vpu, the address of framebuffer
was passed by other driver, like GPU in android, VPU will use
this address to decode to do direction render to archive perfomance.

But this address maybe not a valid phy address, if this happens,
VPU will access this address, and system will hang.

This patch is adding a protection for this, since VPU only
can access phy address, add this check before start using it.

the check is in two phase,
1. if memory address allocated by vmalloc, (kernel vmalloc,
  or user space memory).
2. if the memory in the DDR's address, if it's access beyond
  DDR size, the check will stop.

With this protection, we can replace the hang with a error reported.
so we know the memory given by other(GPU) is wrong,
rather than Hang without any information.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>